+++ pages/_app.tsx
@@ -1,8 +1,7 @@
 import { config } from '@fortawesome/fontawesome-svg-core';
 import '@fortawesome/fontawesome-svg-core/styles.css';
-import { Montserrat } from "next/font/google";
+import { Montserrat } from 'next/font/google';
 import type { AppProps } from 'next/app';
-import Script from 'next/script';
 import '../styles/globals.css';
 
 config.autoAddCss = false;
@@ -18,37 +17,6 @@ export default function App({ Component, pageProps }: AppProps) {
       <div className={montserrat.className}>
         <Component {...pageProps} />
       </div>
-      <Script
-        strategy="beforeInteractive"
-        src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-1638826924479453"
-        async
-        crossOrigin="anonymous"
-      />
-      <Script
-        strategy="beforeInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-          window.dataLayer = window.dataLayer || [];
-          function gtag(){dataLayer.push(arguments);}
-          gtag('consent', 'default', {
-            ad_storage: 'denied',
-            analytics_storage: 'denied',
-          });
-        `,
-        }}
-      />
-      <Script
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-          (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-          new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-          j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-          'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-          })(window,document,'script','dataLayer','GTM-5WC42T9');
-        `,
-        }}
-      />
     </>
   );
 }
